Switch to Qt5 due to Kivi's slowness
Kivy proved itself to be very clumsy and slow, and although it's got a
great language for defining a UI like CSS, it's performace is subpar.
Qt OTOH is faster, and more importantly has "painless" (not really)
integration with matplotlib, which implies numpy, therefore speeding
thing up considerably and significantly.